+++ lib/src/app/controllers/projet_controller.dart
@@ -13,40 +13,41 @@ import 'package:provider/provider.dart';
 ///
 /// Contient le projet actuel dans l'application
 class ProjetController extends ChangeNotifier {
-  /// Projet sélectionné
+  /// Projets
+  List<ProjetModel>? projets;
   ProjetModel? projet;
 
-  /// Liste des événements du projet
+  /// Événements
   List<EvenementModel>? evenements;
-
-  /// Liste des personnages du projet
-  List<PersonnageModel>? personnages;
-
-  /// Liste des lieux du projet
-  List<LieuModel>? lieux;
-
-  /// Liste des objets du projet
-  List<ObjetModel>? objets;
-
-  /// Événement sélectionné
   EvenementModel? evenement;
 
-  /// Personnage sélectionné
+  /// Personnages
+  List<PersonnageModel>? personnages;
   PersonnageModel? personnage;
 
-  /// Lieux sélectionné
+  /// Lieux
+  List<LieuModel>? lieux;
   LieuModel? lieu;
 
-  /// Objet sélectionné
+  /// Objets
+  List<ObjetModel>? objets;
   ObjetModel? objet;
 
   Future _actualiserProjet(ProjetModel projet) async {
     this.projet = projet;
-    await _actualiserListe();
+    await _actualiserApplications();
     notifyListeners();
   }
 
-  Future _actualiserListe() async {
+  Future _chargerProjets() async {
+    projets = [];
+    await FirebaseGlobalTool.recupererListe(ProjetModel.nomCollection, (data) {
+      projets!.add(ProjetModel.fromMap(data));
+    });
+    notifyListeners();
+  }
+
+  Future _actualiserApplications() async {
     evenements = [];
     personnages = [];
     lieux = [];
@@ -116,6 +117,10 @@ class ProjetController extends ChangeNotifier {
     await Provider.of<ProjetController>(context, listen: false)._actualiserProjet(
       Provider.of<ProjetController>(context, listen: false).projet!,
     );
+  }
+
+  static Future chargerProjets(BuildContext context) async {
+    await Provider.of<ProjetController>(context, listen: false)._chargerProjets();
   }
 
   static Future changerEvenement(BuildContext context, String evenementID) async {