+++ changelog.md
@@ -1,3 +1,24 @@
+## [1.0.3]() (2022-11-18)
+
+### Moteur
+
+* Récupération des données liés aux projets lors de leur sélection
+* Gestion des données par projet 
+* Gestion de son profil
+
+# Interface
+
+* Ajout des applications dans l'accueil en sélectionnant un projet
+* Ajout des interfaces (listes, vues, éditions et créations) des événements, personnages et lieux
+* Ajout de la modification de son profil dans les options
+* Mise en couleur de la navigation selon la position dans l'application
+
+## Applications
+
+* L'application événement permet de listé chronologiquement l'histoire
+* L'application personnage permet de créer des personnages liés à l'histoire
+* L'application lieux permet de créer des lieux liés à l'histoire
+
 ## [1.0.2]() (2022-11-01)
 
 ### Moteur